Stopping before an API change that requires changes in dependencies.

+++ modules/juce_core/network/juce_IPAddress.cpp
@@ -166,7 +166,7 @@ IPAddress::IPAddress (const String& adr)
             }
 
             IPAddressByteUnion temp;
-            temp.combined = (uint16) CharacterFunctions::HexParser<int>::parse (tokens[i].getCharPointer());
+            temp.combined = CharacterFunctions::HexParser<uint16>::parse (tokens[i].getCharPointer());
 
             address[i * 2]     = temp.split[0];
             address[i * 2 + 1] = temp.split[1];
